Updates to logo and button
Now when the user clicks on Book now, it sends them to the ticket page, while clicking on the logo will send them to the home page, no matter what page the user is on. Furthermore, when the title has been clicked on, there will be no change in color so it won't turn blue.